debugg

+++ zebra/interface.c
@@ -2560,7 +2560,7 @@ DEFUN (link_params_iscd_scsi_fixed_grid,
 	int i,j,idx_number =1;
 	char grid_fixe[12], grid_fixe1[19];
 	u_int8_t pri=0xFF;
-	u_int8_t cs;
+	u_int16_t cs;
 	u_int8_t grid=1;
 	int16_t n=14; //{30,.....,-14}
 	u_int8_t bitmap[SIZE_BITMAP_TAB];
